+++ services/121-service/src/payments/fsp-integration/intersolve-visa/intersolve-visa-api.error.ts
@@ -1,0 +1,8 @@
+// Usage: throw new IntersolveVisaTransferError('Error message');
+export class IntersolveVisaApiError extends Error {
+  constructor(message?: string) {
+    super(message);
+    this.name = 'IntersolveVisaTransferError';
+    Object.setPrototypeOf(this, new.target.prototype);
+  }
+}
